+++ tests/core.scm
@@ -23,3 +23,26 @@
                                " "
                                (x.foo.toUpperCase))
                 "HEY JO"))))
+
+(test "timing test"
+      (lambda (t)
+        (--> t (is (function? (.. Date.now)) true))
+        (define start (--> Date (now)))
+        (wait 100 (--> t (is (>= (- (--> Date (now)) start) 100) true)))))
+
+(test "values"
+      (lambda (t)
+        (t.is (call-with-values * -) -1)
+        (t.is (call-with-values (lambda () (values 4 5))
+                (lambda (a b) b)) 5)
+        (t.is (call-with-values (lambda () (values 4 5)) +) 9)))
+
+(test "symbols"
+      (lambda (t)
+        (t.is '|foo\x20;bar| (string->symbol "foo bar"))
+        (t.is '|\n| (string->symbol "\n"))
+        (t.is '|\t\t| (string->symbol "\t\t"))
+        (t.is '|\r| (string->symbol "\r"))
+        (t.is '|\s| 's)
+        (t.is '|\x3BB;| 'λ)
+        (t.is '|\x9;\x9;| '|\t\t|)))