+++ tests/std.scm
@@ -1,0 +1,14 @@
+(test "std: case-lambda"
+      (lambda (t)
+
+        ;; example from R7RS without do macro
+        (define range
+          (case-lambda
+           ((e) (range 0 e))
+           ((b e) (let iter ((result ()) (i (- e 1)))
+                    (if (< i b)
+                        result
+                        (iter (cons i result) (- i 1)))))))
+
+        (t.is (range 3) '(0 1 2))
+        (t.is (range 3 5) '(3 4))))