+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * new function `make-list`
 * `(/ 0.0 0.0)` return NaN
 * bookmarklet always fetch latest version
+* `case-lambda` hygienic macro (code from R7RS)
 ### Bugfix
 * allow to use undefined in let
 * fix function call that return NaN